ENH: relocate utility functions and add summary generator
