+++ irrad_control/utils/utils.py
@@ -116,19 +116,3 @@ class Lock:
     def __exit__(self):
         fcntl.lockf(self.lfh.fileno(), fcntl.LOCK_UN)
         self.lfh.close()
-
-
-def duration_str_from_secs(seconds, as_tuple=False):
-
-    days = seconds / (24 * 3600)
-    hours = (days % 1) * 24
-    minutes = (hours % 1) * 60
-    seconds = (minutes % 1) * 60
-    
-    # Return tuple in full days, hours, minutes and seconds
-    res = tuple(int(x) for x in [days, hours, minutes, seconds])
-
-    if as_tuple:
-        return res
-    else:
-        return ", ".join(f"{a[0]}{a[1]}" for a in zip(res, 'dhms') if a[0]) or '0s'
